Initial commit
I've made and finished coding this game before this repository was made. This repository is only a showcase of what I do in the university, as I find this particular project was interesting and I had a lot of fun making it. As you can see, I am the only member of this repository and there is no other contributor, so a github repostitory wasn't really necessary in the first place of making this game